Rename e2e to kindtest
