+++ kindtest/helper_test.go
@@ -1,4 +1,4 @@
-package e2e
+package kindtest
 
 import (
 	"bytes"
