+++ kindtest/manifests/overlays/manager.yaml
